requirements.txt
dice==3.1.2
Faker==9.8.4
PyYAML==6.0
